Naming / Numeric issues

+++ src/net/slipcor/pvparena/arena/Arena.java
@@ -945,8 +945,8 @@ public class Arena {
             final double ratio = getArenaConfig().getDouble(CFG.READY_NEEDEDRATIO);
             getDebugger().i("ratio: " + ratio);
             if (ratio > 0) {
-                final double aRatio = (float) readyPlayers
-                        / (float) players;
+                final double aRatio = readyPlayers
+                        / players;
                 if (players > 0 && aRatio >= ratio) {
                     return "";
                 }
